backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
numpy==2.1.1
scikit-learn==1.5.2
requests==2.32.3
rdkit-pypi==2022.9.5
biopython==1.84
pytest==8.3.3
httpx==0.27.2
